backend/requirements.txt
fastapi
uvicorn
google-ads
pydantic
pydantic-settings
python-dotenv
pandas
